pyproject.toml
[project]
name = "yt-pl-downloader"
version = "0.2.0"
description = "A CLI tool to download YouTube videos and playlists"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "inquirer>=3.4.1",
    "poethepoet>=0.45.0",
    "pyyaml>=6.0.3",
    "yt-dlp>=2025.3.31",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.12.0",
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
